requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
ecdsa==0.17.0
fastapi==0.73.0
h11==0.13.0
httptools==0.3.0
idna==3.3
mongoengine==0.23.1
passlib==1.7.4
pika==1.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pymongo==3.12.3
python-dotenv==0.19.2
python-jose==3.3.0
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.0.post1
uvloop==0.16.0
watchgod==0.7
websockets==10.1
zstandard==0.17.0
